Re adding refactored textures.
If you didnt see the issues i had with the other thermal mods. Basically
if you change the name of a file from something like Servo_0 to servo_0
git dose not recognize the change and basically just breaks.

There is a config option that is supposed to fix this but it dose not.