Models for items, RenderLayers and Transparent Textures.
Fixed TransparentTextures for ducts.
Moved all ducts to use BlockRenderLayer instead of broken int
renderPass.
Attachments use BlockRenderLayer to define what render layer to be on,
This should work for covers that need to be on a different render layer,
Leaves anyone?
Simple Json models for Servos / Filters and such.

On the chopping block next is Covers.
Thaughts: Generate CCModels for each type of cover, then use planar
transforms when rendering to apply the cover to the correct side of the
block.